hoshino/modules/clanbattlereport/requirements.txt
pandas
matplotlib
pillow